docs: better instructions for new langauges
