+++ xtask/codegen/src/js_kinds_src.rs
@@ -1,14 +1,10 @@
 //! Definitions for the ECMAScript AST used for codegen
 //! Based on the rust analyzer parser and ast definitions
 
-use crate::css_kinds_src::CSS_KINDS_SRC;
-use crate::grit_kinds_src::GRIT_KINDS_SRC;
-use crate::html_kinds_src::HTML_KINDS_SRC;
-use crate::json_kinds_src::JSON_KINDS_SRC;
 use crate::kind_src::{KindsSrc, LANGUAGE_PREFIXES};
-use crate::LanguageKind;
 use quote::format_ident;
 use std::collections::BTreeMap;
+use crate::language_kind::LanguageKind;
 
 pub const JS_KINDS_SRC: KindsSrc = KindsSrc {
     punct: &[
@@ -688,13 +684,7 @@ impl Field {
                     _ => name,
                 };
 
-                let kind_source = match language_kind {
-                    LanguageKind::Js => JS_KINDS_SRC,
-                    LanguageKind::Css => CSS_KINDS_SRC,
-                    LanguageKind::Json => JSON_KINDS_SRC,
-                    LanguageKind::Grit => GRIT_KINDS_SRC,
-                    LanguageKind::Html => HTML_KINDS_SRC,
-                };
+                let kind_source = language_kind.to_kinds();
 
                 // we need to replace "-" with "_" for the keywords
                 // e.g. we have `color-profile` in css but it's an invalid ident in rust code
